Rename allowedFighterSubtypes to fighter-subtype-picker
The module now holds both halves of the subtype picker — which subtypes a form
may offer, and how ticking one updates the selection — so a name describing only
the allowlist undersold it. Scope is picking only; ranking stays in
fighterSubtypeRank.ts and subtype-driven effects in fighter-subtype-grants.ts,
which the new module header spells out so the four subtype files in utils/ stay
tellable apart.

The two imports in custom-fighters.tsx collapse into one now that both helpers
share a module.
