+++ components/admin/admin-create-fighter-type.tsx
@@ -10,7 +10,7 @@ import { HiX } from "react-icons/hi";
 import { GangType, Equipment } from "@/types/gang";
 import { EditionSelect, useEditions } from '@/components/edition-select';
 import { hasAlignment, hasSaveCharacteristic, allowsMultipleSubtypes, hasStartingXp, hasVehicles } from '@/types/edition';
-import { toggleFighterSubtype } from '@/utils/allowedFighterSubtypes';
+import { toggleFighterSubtype } from '@/utils/fighter-subtype-picker';
 import { getSkillSetRank } from "@/utils/skillSetRank";
 import { compareEquipmentCategories } from "@/utils/getEquipmentCategoryRank";
